requirements.txt
# === Основні бібліотеки ===
openai>=1.3.7
faiss-cpu>=1.7.4
telethon>=1.33.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0

# === FastAPI + сервер ===
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# === Шаблонізація + веб ===
jinja2>=3.1.2
aiofiles>=23.2.1  # Для обслуговування статики (опціонально)

# === Аналіз/класифікація (опціонально) ===
scikit-learn>=1.3.0
